Move EventsHeader to Organisms
