--- src/EventsHeader/EventsHeaderButton.module.css
@@ -1,13 +1,0 @@
-.button {
-	display: inline-block;
-	padding: 8px;
-	margin: 5px;
-	background-color: white;
-	width: 100px;
-	height: 100px;
-}
-
-.box {
-	height: 100%;
-	width: 100%;
-}
