+++ src/components/Molecules/List/Events/Events.tsx
@@ -2,7 +2,7 @@ import React, { SFC } from 'react';
 import Container from '../../../../Container/Container';
 import Section from '../../../../Section/Section';
 import { translate, TranslationFunction } from 'react-i18next';
-import EventsHeader from '../../../../EventsHeader/EventsHeader';
+import EventsHeader from '../../../Organisms/EventsHeader/EventsHeader';
 import EventCard from '../../../../EventCard/EventCard';
 import compose from 'recompose/compose';
 import { IMeetupEvent } from '../../../../modules/EventsContainer/EventsContainer.interfaces';
