+++ src/components/Organisms/EventsHeader/EventsHeader.module.scss
@@ -1,4 +1,4 @@
-@import '../theming/config.module';
+@import '../../../theming/config.module';
 
 .nextEventTitle {
 	color: white;
